refactor: split into two demos and extract the data preprocessing logic into 'util'.
